Streamlit-failure-analysis-metal-classifier-2/requirements.txt
tensorflow==2.2.0
streamlit==0.52.0
numpy==1.16.4
pillow
keras